% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccrm_est.R
\name{ccrm_est}
\alias{ccrm_est}
\title{The wrapper of estimation functions}
\usage{
ccrm_est(x, y, z, s_max = 4, theta_init = NULL, est_method = "gmm")
}
\arguments{
\item{x}{regressor (N-by-1)}

\item{y}{dependent variable (N-by-1)}

\item{z}{control varibles (N-by-p)}

\item{theta_init}{initial value for distribution parameters}

\item{est_method}{}
}
\value{
A list contains estimated coefficients and inferential statistics
\item{theta_b}{Estimated distributional parameter p, b_L, b_H}
\item{theta_m}{Estimated moments of beta_i}
\item{theta_u}{Estimated moments of u_i}
\item{gamma}{estimated coefficients of control varibles}
\item{V_theta}{variance}
}
\description{
The wrapper of estimation functions
}
