% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func_K2.R
\name{ccrm_est_2}
\alias{ccrm_est_2}
\title{GMM Estimator of distributional parameters of a K = 2 random coefficient model}
\usage{
ccrm_est_2(
  x,
  y,
  z = NULL,
  theta_init = NULL,
  s_max = 4,
  remove_intercept = TRUE,
  iter_gmm = TRUE,
  seed = 2023
)
}
\arguments{
\item{x}{regressor (N-by-1)}

\item{y}{dependent variable (N-by-p_x)}

\item{z}{control variables (N-by-p_z)}

\item{theta_init}{Initial estimates of distributional parameters}

\item{s_max}{Maximum order of moments used in estimation}

\item{remove_intercept}{whether subtract estimated intercept term in calculation of y_tilde}

\item{iter_gmm}{Whether to use iterative GMM (If FALSE, use OW-GMM with initial estimates)}

\item{seed}{seed for random number generator}
}
\value{
A list contains estimated coefficients and inferential statistics
\item{theta_b}{Estimated distributional parameter p, b_L, b_H}
\item{theta_b_se}{s.e. of Estimated distributional parameter p, b_L, b_H}
\item{theta_m}
\item{theta_m_se}{s.e. of estimated moments of beta_i}
\item{theta_m_gmm}{Estimated moments of beta_i from gmm}
\item{theta_m_gmm_se}{s.e. estimated moments of beta_i from gmm}
\item{gamma}{estimated coefficients of control varibles, including intercept}
\item{gamma_se}{s.e. estimated coefficients of control varibles, including intercept}
\item{V_theta}{variance}
\item{gmm_res}{GMM estimation of moments output object}
}
\description{
This function estimates the distribution of beta_i using GMM. The moment conditions are
listed in Section 3 of Gao and Pesaran (2023)
}
