% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func_hetero_K2.R
\name{ccrm_est_hetero}
\alias{ccrm_est_hetero}
\title{Estimation: heteroskedastic error, two categories}
\usage{
ccrm_est_hetero(x, y, z, s_max, theta_init = NULL)
}
\arguments{
\item{x}{regressors (N-by-1)}

\item{y}{dependent variable (N-by-1)}

\item{s_max}{maximum order of moments of x}

\item{theta_init}{initial value for distribution parameters}
}
\value{
A list contains estimated coefficients and inferential statistics
\item{theta_b}{Estimated distributional parameter p, b_L, b_H}
\item{theta_m}{Estimated moments of beta_i}
\item{theta_u}{Estimated moments of u_i}
\item{gamma}{estimated coefficients of control varibles}
\item{V_theta}{variance}
}
\description{
Estimation: heteroskedastic error, two categories
}
