% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func_K2.R
\name{ccrm_est_K2}
\alias{ccrm_est_K2}
\title{Estimation: two categories}
\usage{
ccrm_est_K2(
  x,
  y,
  z,
  theta_init = NULL,
  s_max = 4,
  weight_mat = NULL,
  iter_gmm = TRUE
)
}
\arguments{
\item{x}{regressor (N-by-1)}

\item{y}{dependent variable (N-by-1)}

\item{theta_init}{initial value for distribution parameters}

\item{iter_gmm}{}
}
\value{
A list contains estimated coefficients and inferential statistics
\item{theta_b}{Estimated distributional parameter p, b_L, b_H}
\item{theta_m}{Estimated moments of beta_i}
\item{theta_u}{Estimated moments of u_i}
\item{gamma}{estimated coefficients of control varibles, including intercept}
\item{V_theta}{variance}
}
\description{
Estimation: two categories
}
