% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func_K2_sym.R
\name{ccrm_est_K2_sym}
\alias{ccrm_est_K2_sym}
\title{Estimation: two categories, symmetric error}
\usage{
ccrm_est_K2_sym(x, y, z, theta_init = NULL, s_max = 4)
}
\arguments{
\item{x}{regressor (N-by-1)}

\item{y}{dependent variable (N-by-1)}

\item{z}{control varibles (N-by-p)}

\item{theta_init}{initial value for distribution parameters}

\item{s_max}{}
}
\value{
A list contains estimated coefficients and inferential statistics
\item{theta_b}{Estimated distributional parameter p, b_L, b_H}
\item{theta_m}{Estimated moments of beta_i}
\item{theta_u}{Estimated moments of u_i}
\item{gamma}{estimated coefficients of control varibles}
\item{V_theta}{variance}
}
\description{
Estimation: two categories, symmetric error
}
