% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chip_data}
\alias{chip_data}
\title{China Household Income Project (CHIP) data (subsample)}
\format{
\subsection{\code{chip_data}}{

A data frame with 17,371 rows and 10 columns:
\describe{
\item{edu_group}{education level: 1 = postsecondary; 0 = high school or less}
\item{log_wage}{log of wages (total wage income / hours)}
\item{edu_year}{years to education}
...
}
}
}
\usage{
chip_data
}
\description{
A subset of China Household Income Project (CHIP) data
in 2018 with urban households
}
\keyword{datasets}
