% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func_K2.R
\name{moment_est}
\alias{moment_est}
\title{Direct moment estimator using the identification equations}
\usage{
moment_est(x, y, z = NULL, remove_intercept = TRUE)
}
\arguments{
\item{x}{regressor (N-by-1)}

\item{y}{dependent variable (N-by-p_x)}

\item{z}{control variables (N-by-p_z)}

\item{remove_intercept}{whether subtract estimated intercept term in calculation of y_tilde}
}
\value{
A list contains
\item{para_est}{estimated (a, sigma2, sigma3, b1, b2, b3) }
\item{remove_intecept}{Save the parameter remove_intercept for future reference. = NA if z is not provided.}
}
\description{
Direct moment estimator using the identification equations
}
