% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func_K3.R
\name{moment_est_gmm_3}
\alias{moment_est_gmm_3}
\title{Estimation of moments of beta_i}
\usage{
moment_est_gmm_3(x, y, z, s_max = 6, second_step = TRUE)
}
\arguments{
\item{second_step}{}
}
\description{
Estimation of moments of beta_i
}
