% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func_K2_binary.R
\name{moment_est_gmm_binary}
\alias{moment_est_gmm_binary}
\title{Estimation of moments of beta_i}
\usage{
moment_est_gmm_binary(x, y, z, s_max, second_step = TRUE)
}
\arguments{
\item{second_step}{}
}
\description{
Estimation of moments of beta_i
}
