% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func_K2_sym.R
\name{moment_est_gmm_sym}
\alias{moment_est_gmm_sym}
\title{Estimation of moments of beta_i with symmetric error}
\usage{
moment_est_gmm_sym(x, y, s_max = 4)
}
\arguments{
\item{s_max}{}
}
\description{
Estimation of moments of beta_i with symmetric error
}
