% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func_hetero_K2.R
\name{moment_est_md}
\alias{moment_est_md}
\title{First step estimation minimum distance}
\usage{
moment_est_md(x, y, s_max)
}
\arguments{
\item{s_max}{maximum order of moments of x}
}
\value{
theta_hat (a, b1, b2, b3)
}
\description{
First step estimation minimum distance
}
