% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_est.R
\name{phi_est}
\alias{phi_est}
\title{Estimation: gamma}
\usage{
phi_est(x, y, z)
}
\arguments{
\item{x}{regressor (N-by-1)}

\item{y}{dependent variable (N-by-1)}

\item{z}{control variables (N-by-p_z)}
}
\value{
A list contains estimated coefficients and inferential statistics
\item{phi_hat}{Estimated phi}
\item{se_phi}{Estimated  se}
}
\description{
Estimation: gamma
}
