% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func.R
\name{ccrm_est}
\alias{ccrm_est}
\title{Estimation: homoskedastic error, two categories}
\usage{
ccrm_est(x, y, theta_init = NULL, s_1 = 3, s_2 = 2, s_3 = 1)
}
\arguments{
\item{x}{regressor (N-by-1)}

\item{y}{dependent variable (N-by-1)}

\item{theta_init}{initial value for distribution parameters}

\item{s_2}{}

\item{s_3}{}
}
\value{
A list contains estimated coefficients and inferential statistics
\item{theta_hat}{Estimated coefficient}
\item{wald_stat}{Wald statistic}
\item{t_stat}{t statistic}
\item{V_theta}{variance}
}
\description{
Estimation: homoskedastic error, two categories
}
