% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func_K2.R, R/est_func_K3.R
\name{moment_est_gmm}
\alias{moment_est_gmm}
\title{Estimation of moments of beta_i}
\usage{
moment_est_gmm(x, y, z, s_max)

moment_est_gmm(x, y, z, s_max)
}
\arguments{
\item{s_max}{}
}
\description{
Estimation of moments of beta_i

Estimation of moments of beta_i
}
